src/backend/requirements.txt
# --- Backend & API Server ---
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.0

# --- LangChain Core & Orchestration ---
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-experimental>=0.0.59

# --- LLM Integrations (Cerebros) ---
langchain-groq>=0.1.3        
langchain-google-genai>=1.0.3 

# --- RAG & Vector Store (Memoria) ---
faiss-cpu>=1.8.0              # Base de datos vectorial local (Si tienes GPU usa faiss-gpu)
langchain-text-splitters>=0.2.0

# --- Embeddings (Opciones) ---
# Si usas 'text-embedding-3-small' (como en tu server.py anterior):
langchain-openai>=0.1.6Microsoft.QuickAction.Bluetooth
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.6
# Si prefieres usar embeddings de Google (Gratis en tier bajo):
# (Ya incluido en langchain-google-genai)

# --- Tools & Data Handling ---
pandas>=2.2.0                 # Para leer los CSVs de Kaggle
kagglehub>=0.2.4              # Para la tool kaggle_search
sqlalchemy>=2.0.29            # Para ejecutar SQL sobre los datos
tabulate>=0.9.0               # Formateo de tablas para que el LLM las lea bien
numexpr>=2.10.0               # Dependencia común de pandas para cálculos
wikipedia-api